Moved ganache related code to its own file
