+++ src/tasks/start/utils/backend/backend.ts
@@ -1,6 +1,4 @@
-import { promisify } from 'util'
 import chokidar from 'chokidar'
-import { BuidlerPluginError } from '@nomiclabs/buidler/plugins'
 import { BuidlerRuntimeEnvironment } from '@nomiclabs/buidler/types'
 import { createDao } from './dao'
 import { deployImplementation } from './app'
@@ -12,20 +10,8 @@ import { logBack } from '../logger'
 import { readArapp } from '../arapp'
 import { AragonConfig, AragonConfigHooks } from '~/src/types'
 import { TASK_COMPILE } from '../../../task-names'
-import tcpPortUsed from 'tcp-port-used'
 import deployAragonBases from './deployAragonBases'
-import {
-  aragenGasLimit,
-  aragenMnemonic,
-  testnetPort
-} from '../../../../aragon-params'
-// There's an issue with how web3 exports its typings that conflicts with
-// ganache-core imports of those typings. Follow https://github.com/trufflesuite/ganache-core/issues/465
-// for upcoming solutions, meanwhile require is used to ignore the types
-/* eslint-disable @typescript-eslint/no-var-requires */
-const ganache = require('ganache-core')
-
-const buidlerevmNetworkName = 'buidlerevm'
+import { startGanache } from './ganache'
 
 /**
  * Starts the task's backend sub-tasks. Logic is contained in ./tasks/start/utils/backend/.
@@ -46,28 +32,10 @@ export async function startBackend(
   await bre.run(TASK_COMPILE)
 
   /**
-   * Until BuidlerEVM JSON RPC is ready a ganache server will be started
+   * Until BuidlerEVM JSON RPC is ready, a ganache server will be started
    * on the appropiate conditions
    */
-  if (bre.network.name === buidlerevmNetworkName) {
-    throw new BuidlerPluginError(
-      `Cannot use ${buidlerevmNetworkName} network for this task until a JSON RPC is exposed`
-    )
-  } else if (bre.network.name === 'localhost') {
-    if (await tcpPortUsed.check(testnetPort)) {
-      logBack(`Connecting to existing local blockchain instance`)
-    } else {
-      logBack(`Starting a new Ganache testnet instance`)
-      const server = ganache.server({
-        gasLimit: aragenGasLimit,
-        mnemonic: aragenMnemonic
-      })
-      const blockchain = await promisify(server.listen)(testnetPort)
-      logBack(
-        `New Ganache instance ready, id: ${blockchain.options.network_id}`
-      )
-    }
-  }
+  await startGanache(bre)
 
   // Deploy bases
   const { ensAddress, daoFactoryAddress, apmAddress } = await deployAragonBases(
